# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reference App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /references:
    get:
      operationId: listReferences
      tags: [references]
      summary: List all references
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of references
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reference"
    post:
      operationId: createReference
      tags: [references]
      summary: Create a reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferenceBody"
      responses:
        "201":
          description: Reference created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /references/{id}:
    get:
      operationId: getReference
      tags: [references]
      summary: Get a reference by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateReference
      tags: [references]
      summary: Update a reference
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReferenceBody"
      responses:
        "200":
          description: Updated reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reference"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /references/{referenceId}/nodes:
    get:
      operationId: listReferenceNodes
      tags: [nodes]
      summary: List nodes for a reference
      parameters:
        - name: referenceId
          in: path
          required: true
          schema:
            type: integer
        - name: parentNodeId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of reference nodes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferenceNode"
    post:
      operationId: createReferenceNode
      tags: [nodes]
      summary: Create a reference node
      parameters:
        - name: referenceId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReferenceNodeBody"
      responses:
        "201":
          description: Reference node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceNode"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes
      parameters:
        - name: referenceNodeId
          in: query
          required: false
          schema:
            type: integer
        - name: visibility
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Note deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - createdAt
...
    Reference:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [BIBLE, MUSIC, POEM, BOOK, SERMON]
        title:
          type: string
        abbreviation:
          type: ["string", "null"]
        author:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - createdAt
    CreateReferenceBody:
      type: object
      properties:
        type:
          type: string
          enum: [BIBLE, MUSIC, POEM, BOOK, SERMON]
        title:
          type: string
        abbreviation:
          type: string
        author:
          type: string
        description:
          type: string
      required:
        - type
        - title
    UpdateReferenceBody:
      type: object
      properties:
        type:
          type: string
          enum: [BIBLE, MUSIC, POEM, BOOK, SERMON]
        title:
          type: string
        abbreviation:
          type: string
        author:
          type: string
        description:
          type: string
    ReferenceNode:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [BOOK, CHAPTER, VERSE, PAGE, SESSION, STANZA, LINE, PARAGRAPH]
        content:
          type: ["string", "null"]
        label:
          type: string
        referenceId:
          type: integer
        parentNodeId:
          type: ["integer", "null"]
        position:
          type: integer
      required:
        - id
        - type
        - label
        - referenceId
        - position
    CreateReferenceNodeBody:
      type: object
      properties:
        type:
          type: string
          enum: [BOOK, CHAPTER, VERSE, PAGE, SESSION, STANZA, LINE, PARAGRAPH]
        content:
          type: string
        label:
          type: string
        parentNodeId:
          type: ["integer", "null"]
        position:
          type: integer
      required:
        - type
        - label
        - position
    Note:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        content:
          type: string
        referenceNodeId:
          type: ["integer", "null"]
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC, CHANNEL]
        channelId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        referenceNode:
          type: object
          nullable: true
          properties:
            id:
              type: integer
            label:
              type: string
            referenceId:
              type: integer
          required:
            - id
            - label
            - referenceId
      required:
        - id
        - userId
        - content
        - visibility
        - createdAt
        - user
    CreateNoteBody:
      type: object
      properties:
        content:
          type: string
        referenceNodeId:
          type: ["integer", "null"]
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC, CHANNEL]
        channelId:
          type: ["integer", "null"]
      required:
        - content
        - visibility
    UpdateNoteBody:
      type: object
      properties:
        content:
          type: string
        visibility:
          type: string
          enum: [PRIVATE, PUBLIC, CHANNEL]
        channelId:
          type: ["integer", "null"]
...