# lib/api-zod/src/generated/types/createNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reference App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateNoteBodyVisibility } from "./createNoteBodyVisibility";

export interface CreateNoteBody {
  content: string;
  /** @nullable */
  referenceNodeId?: number | null;
  visibility: CreateNoteBodyVisibility;
  /** @nullable */
  channelId?: number | null;
}