# lib/api-zod/src/generated/types/createReferenceBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reference App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReferenceBodyType } from "./createReferenceBodyType";

export interface CreateReferenceBody {
  type: CreateReferenceBodyType;
  title: string;
  abbreviation?: string;
  author?: string;
  description?: string;
}