# lib/api-zod/src/generated/types/referenceNode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reference App API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReferenceNodeType } from "./referenceNodeType";

export interface ReferenceNode {
  id: number;
  type: ReferenceNodeType;
  /** @nullable */
  content?: string | null;
  label: string;
  referenceId: number;
  /** @nullable */
  parentNodeId?: number | null;
  position: number;
}